pyproject.toml
[tool.poetry]
name = "checklist"
version = "0.1.0"
description = ""
authors = ["Alisher007 <alisherx07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.0"
django-allauth = "^0.52.0"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
dj-rest-auth = "^2.2.7"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
